pyproject.toml
[project]
name = "fashion_trends"
version = "0.1.0"
description = ""
authors = [
    {name = "Chaitanya patakota"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "tweepy (>=4.15.0,<5.0.0)",
    "streamlit (>=1.42.2,<2.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "pyngrok (>=7.2.3,<8.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "asyncpraw (>=7.8.1,<8.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
